changelog 1.1.31

+++ app/api/embarcacao/delete/route.ts
@@ -1,7 +1,13 @@
-import prisma from '@/lib/prisma';
+import { validateRequest } from "@/auth";
+import prisma from "@/lib/prisma";
 
 export async function DELETE(req: Request) {
   const { id } = await req.json();
+  const { user } = await validateRequest();
+
+  if (!user) {
+    return new Response("Unauthorized", { status: 401 });
+  }
 
   const result = await prisma.embarcacao.delete({
     where: {
